feat: add more readme
