+++ samples/ONNX2TRT_opengpu/onnx2trt_app/packages/201125699002-onnx2trt_app-1.0/descriptor.json
@@ -1,0 +1,11 @@
+{
+    "runtimeDescriptor":
+    {
+        "envelopeVersion": "2021-01-01",
+        "entry":
+        {
+            "path": "python3.6",
+            "name": "/panorama/wrapper-app.py"
+        }
+    }
+}
