+++ samples/ONNX2TRT_opengpu/onnx2trt_app/packages/201125699002-onnx2trt_app-1.0/src/.gitignore
@@ -1,0 +1,2 @@
+# pycache
+__pycache__/
