+++ samples/ONNX2TRT_opengpu/onnx2trt_app/packages/201125699002-onnx2trt_app-1.0/src/test/.gitignore
@@ -1,0 +1,3 @@
+# Engine file
+*.engine
+res-*.jpg
