--- samples/ONNX2TRT_opengpu/onnx2trt_app/packages/201125699002-ppe_panorama_app-1.0/descriptor.json
@@ -1,11 +1,0 @@
-{
-    "runtimeDescriptor":
-    {
-        "envelopeVersion": "2021-01-01",
-        "entry":
-        {
-            "path": "python3.6",
-            "name": "/panorama/wrapper-app.py"
-        }
-    }
-}
