--- samples/ONNX2TRT_opengpu/onnx2trt_app/packages/201125699002-ppe_panorama_app-1.0/package.json
@@ -1,51 +1,0 @@
-{
-    "nodePackage": {
-        "envelopeVersion": "2021-01-01",
-        "name": "ppe_panorama_app",
-        "version": "1.0",
-        "description": "Default description for package ppe_panorama_app",
-        "assets": [
-            {
-                "name": "ppe_panorama_app",
-                "implementations": [
-                    {
-                        "type": "container",
-                        "assetUri": "7a80d9b0ae5c3697d201eef3e891fc53f55c31be3a7c4c4fc7b130d1c4d46246.tar.gz",
-                        "descriptorUri": "63a87b807a1a19dea23660a2e2ca168a9377b6dca80fb1bfb3e9e123702f26a8.json",
-                        "requirements": 
-                        [{
-                            "type" : "hardware_access",
-                            "inferenceAccelerators": [ 
-                                {
-                                    "deviceType": "nvhost_gpu",
-                                    "sharedResourcePolicy": {
-                                        "policy" : "allow_all"
-                                    }
-                                }
-                            ]
-                        }]
-                    }
-                ]
-            }
-        ],
-        "interfaces": [
-            {
-                "name": "ppe_panorama_app_interface",
-                "category": "business_logic",
-                "asset": "ppe_panorama_app",
-                "inputs": [
-                    {
-                        "name": "video_in",
-                        "type": "media"
-                    }
-                ],
-                "outputs": [
-                    {
-                        "name": "video_out",
-                        "type": "media"
-                    }
-                ]
-            }
-        ]
-    }
-}
