--- samples/ONNX2TRT_opengpu/onnx2trt_app/packages/201125699002-ppe_panorama_app-1.0/src/test/.gitignore
@@ -1,3 +1,0 @@
-# Engine file
-*.engine
-res-*.jpg
